Don't use sentinel terminated arrays
Zig's type system requires that there is *always* a 0 at the end of the
array if the array is sentinel-terminated. However, C's type system is
not strong enough to ensure this. Currently wlroots zero-initializes
most if not all structs but this behavior may change in the future.

Instead, just use normal arrays and rely on std.mem.sliceTo().