Use a recursive refAllDecls implementation
This saves us the boiler plate of adding a test block with a
testing.refAllDecls() call to every container type in zig-wlroots.

This is restricted to recursing on pub decls to avoid recursing into
dependencies and thes std, which is fine for the purposes of zig-wlroots
as everything is pub.